+++ velox/dwio/dwrf/reader/ReaderBase.h
@@ -101,7 +101,7 @@ class ReaderBase {
         schema_{std::dynamic_pointer_cast<const RowType>(convertType(*footer))},
         fileLength_{0},
         psLength_{0} {
-    DWIO_ENSURE(footer_->getDwrfFooter()->GetArena());
+    DWIO_ENSURE(footer_->getDwrfPtr()->GetArena());
     DWIO_ENSURE_NOT_NULL(schema_, "invalid schema");
     if (!handler_) {
       handler_ = encryption::DecryptionHandler::create(*footer);
